Made Shirt Size Pop-Up
The pop up with the dismissable modal has been created. The user can successfully choose exactly one shirt size. There are still some issues such as the fact that they can just click out of the modal or that they can submit it while empty but that can be fixed. Also, the users choice isn't being sent anywhere at this time, but it is being added to their private data. Also added a tshirt size to the users private data and added a const to set the users shirt size in firebaseUtils.ts